requirements.txt
serial
tornado
paho-mqtt
